+++ extensions/bundesliga/src/utils/index.ts
@@ -1,0 +1,26 @@
+export const getFlagEmoji = (isoCode: string) => {
+  if (isoCode === "GB-ENG") {
+    return "🏴󠁧󠁢󠁥󠁮󠁧󠁿";
+  }
+  if (isoCode === "GB-WLS") {
+    return "🏴󠁧󠁢󠁷󠁬󠁳󠁿";
+  }
+  if (isoCode === "GB-SCT") {
+    return "🏴󠁧󠁢󠁳󠁣󠁴󠁿";
+  }
+  if (isoCode === "GB-NIR") {
+    // The only official flag in Northern Ireland is the Union Flag of the United Kingdom.
+    return "🇬🇧";
+  }
+
+  return isoCode
+    .toUpperCase()
+    .replace(/./g, (char) => String.fromCodePoint(127397 + char.charCodeAt(0)));
+};
+
+export const positionMap = new Map<string, string>([
+  ["GOALKEEPER", "Goalkeeper"],
+  ["DEFENSE", "Defender"],
+  ["MIDFIELD", "Midfielder"],
+  ["ATTACK", "Striker"],
+]);