+++ extensions/f1-standings/src/views/RaceList.tsx
@@ -1,9 +1,10 @@
 import { useEffect, useState } from "react";
-import { ActionPanel, Action, List, Color } from "@raycast/api";
+import { ActionPanel, Action, List, Color, useNavigation, Icon } from "@raycast/api";
 import { useFormula1RaceUrl, useRaces, useSeasons } from "../hooks";
 import { formatDate, getFlag, getRaceDates } from "../utils";
 import RaceSessionDetails from "../components/RaceSessionDetails";
 import RaceResultList from "../views/RaceResultList";
+import { AddToCalendar } from "./AddToCalendar";
 import { Race } from "../types";
 
 function RaceList() {
@@ -14,6 +15,7 @@ function RaceList() {
   const seasons = useSeasons();
   const [pastRaces, upcomingRaces, isLoading] = useRaces(season);
   const raceUrl = useFormula1RaceUrl(season, selectedRace);
+  const { push } = useNavigation();
 
   useEffect(() => {
     const upcomingRounds = Object.keys(upcomingRaces);
@@ -75,12 +77,13 @@ function RaceList() {
                 actions={
                   <ActionPanel title={race.raceName}>
                     <Action.Push
+                      icon={{ source: "flag-checkered.png", tintColor: Color.Green }}
                       title="Show Results"
                       target={<RaceResultList season={season} round={race.round} />}
-                      icon={{ source: "flag-checkered.png", tintColor: Color.Green }}
                     />
                     {raceDates.length ? (
                       <Action
+                        icon={Icon.Sidebar}
                         title={isShowingDetail ? "Hide Sessions" : "Show Sessions"}
                         onAction={() => setIsShowingDetail((previous) => !previous)}
                       />
@@ -120,10 +123,16 @@ function RaceList() {
                   <ActionPanel title={race.raceName}>
                     {raceDates.length ? (
                       <Action
+                        icon={Icon.Sidebar}
                         title={isShowingDetail ? "Hide Sessions" : "Show Sessions"}
                         onAction={() => setIsShowingDetail((previous) => !previous)}
                       />
                     ) : null}
+                    <Action
+                      title={"Add to calendar"}
+                      icon={Icon.Calendar}
+                      onAction={() => push(<AddToCalendar race={race} raceDates={raceDates} />)}
+                    />
                     <Action.OpenInBrowser
                       title="View on Wikipedia.org"
                       url={race.url || race.Circuit.url}