+++ extensions/twitch/src/following.tsx
@@ -1,37 +1,20 @@
-import {
-  ActionPanel,
-  confirmAlert,
-  Detail,
-  getPreferenceValues,
-  ImageMask,
-  List,
-  ListItem,
-  OpenAction,
-  showToast,
-  ToastStyle,
-} from "@raycast/api";
-import React from "react";
+import { getPreferenceValues, Image, List, showToast, Toast } from "@raycast/api";
+import { useEffect, useState } from "react";
 import fetch from "node-fetch";
-import { exec } from "child_process";
-
 import Item from "./interfaces/FollowingItem";
 import { Preferences } from "./interfaces/Preferences";
-import User from "./interfaces/user";
 import Action from "./utils";
 
 export default function main() {
   const preferences: Preferences = getPreferenceValues();
   const clientId = preferences.clientId;
   const authorization = preferences.authorization;
-  const streamlinkLocation = preferences.streamlink || "/opt/homebrew/bin/streamlink";
-  const quality = preferences.quality || "best";
 
-  const [loading, setLoading] = React.useState(false);
-  const [query, setQuery] = React.useState<string>("");
-  const [userId, setUserId] = React.useState<string>("");
-  const [items, setItems] = React.useState<Item[]>([]);
+  const [loading, setLoading] = useState(true);
+  const [userId, setUserId] = useState<string>("");
+  const [items, setItems] = useState<Item[]>([]);
 
-  React.useEffect(() => {
+  useEffect(() => {
     setLoading(true);
     fetch(`https://api.twitch.tv/helix/users`, {
       headers: {
@@ -44,12 +27,12 @@ export default function main() {
         if (data && data.data) {
           setUserId(data.data[0].id);
         } else if (data.error && data.error.toLowerCase().includes("invalid")) {
-          showToast(ToastStyle.Failure, data.message);
+          showToast(Toast.Style.Failure, data.message);
         }
       });
   }, []);
 
-  React.useEffect(() => {
+  useEffect(() => {
     if (!userId) return;
 
     fetch(`https://api.twitch.tv/helix/streams/followed?user_id=${userId}`, {
@@ -67,36 +50,37 @@ export default function main() {
           (data.error && data.message.toLowerCase().includes("invalid")) ||
           data.message.toLowerCase().includes("missing")
         ) {
-          showToast(ToastStyle.Failure, data.message);
+          showToast(Toast.Style.Failure, data.message);
         }
       });
   }, [userId]);
 
   return (
     <>
-      <List
-        isLoading={loading}
-        searchBarPlaceholder="Search for a Streamer on Twitch"
-        navigationTitle="Search a Channel"
-        onSearchTextChange={(text) => setQuery(text)}
-      >
+      <List isLoading={loading} searchBarPlaceholder="Search for a streamer...">
         {items.map((item: Item) => {
           return (
-            <ListItem
+            <List.Item
               key={item.id}
               icon={{
                 source: item.thumbnail_url.replace("{width}", "320").replace("{height}", "180"),
-                mask: ImageMask.RoundedRectangle,
+                mask: Image.Mask.RoundedRectangle,
               }}
-              accessoryIcon={{
-                tintColor: item.type == "live" ? "green" : "red",
-                source: item.type == "live" ? "checkmark-circle-16" : "xmark-circle-16",
-              }}
-              accessoryTitle={`${item.type == "live" ? item.game_name : "Offline"}`}
+              accessories={[
+                {
+                  text: item.type == "live" ? item.game_name : "Offline",
+                },
+                {
+                  icon: {
+                    source: item.type == "live" ? "checkmark-circle-16" : "xmark-circle-16",
+                    tintColor: item.type == "live" ? "green" : "red",
+                  },
+                },
+              ]}
               id={item.id}
               title={item.title}
               subtitle={item.user_name}
-              actions={<Action live={item.type == "live"} name={item.user_name} />}
+              actions={<Action live={item.type == "live"} name={item.user_login} />}
             />
           );
         })}