+++ extensions/json-format/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
 
-## Added Format JSON (2021-12-10)
-Initial version
+## [Fixes] - 2022-04-20
+- Moved indentation to preferences
+
+## [Added Format JSON] - 2021-12-10
+- Initial version